jira training
svr technologies provides all training IT courses. This self-paced course is for developers and architects who want to get hands-on experience using Anypoint Platform to build APIs and integrations. In the first part, students use Anypoint Platform discover, consume, design, build, deploy, manage, and govern APIs. In the second part, students focus on using Mule and Anypoint Studio to build applications for use as API implementations and/or integrations